feat: refactor score select component with type based props
