+++ packages/html/ds/src/score-select/types.ts
@@ -1,17 +1,14 @@
 import { ButtonSize } from '../button/types';
 
-type ButtonOption = {
-  label: string;
-  value: string;
-};
+type ScoreSelectType = '1-5' | '1-7' | '0-10';
 
 export type ScoreSelectProps = {
   name: string;
+  value?: string;
   size?: ButtonSize;
-  defaultValue?: string;
-  options?: ButtonOption[];
-  label?: string;
+  label: string;
   hint?: string;
   leftLabel?: string;
   rightLabel?: string;
+  type: ScoreSelectType;
 };
