+++ packages/react/ds/src/score-select/type.ts
@@ -1,18 +1,15 @@
 import { ButtonSize } from '../button/types.js';
 
-type Option = {
-  value: string;
-  label: string;
-};
+type ScoreSelectType = '1-5' | '1-7' | '0-10';
 
 export type ScoreSelectProps = {
   name: string;
+  value?: string;
   size?: ButtonSize;
-  options: Option[];
-  defaultValue?: string;
   label: string;
   hint?: string;
   leftLabel?: string;
   rightLabel?: string;
+  type: ScoreSelectType;
   onChange?: (value: string) => void;
 };
